// coffeeclub/src/features/api/api/generated/models/CreateDrinkOrderDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CoffeeClub.Core.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Drink } from './Drink';
import {
    DrinkFromJSON,
    DrinkFromJSONTyped,
    DrinkToJSON,
} from './Drink';
import type { MilkType } from './MilkType';
import {
    MilkTypeFromJSON,
    MilkTypeFromJSONTyped,
    MilkTypeToJSON,
} from './MilkType';

/**
 * 
 * @export
 * @interface CreateDrinkOrderDto
 */
export interface CreateDrinkOrderDto {
    /**
     * 
     * @type {string}
     * @memberof CreateDrinkOrderDto
     */
    coffeeBeanId?: string;
    /**
     * 
     * @type {Drink}
     * @memberof CreateDrinkOrderDto
     */
    drink?: Drink;
    /**
     * 
     * @type {MilkType}
     * @memberof CreateDrinkOrderDto
     */
    milkType?: MilkType;
    /**
     * 
     * @type {boolean}
     * @memberof CreateDrinkOrderDto
     */
    isIced?: boolean;
}

/**
 * Check if a given object implements the CreateDrinkOrderDto interface.
 */
export function instanceOfCreateDrinkOrderDto(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function CreateDrinkOrderDtoFromJSON(json: any): CreateDrinkOrderDto {
    return CreateDrinkOrderDtoFromJSONTyped(json, false);
}

export function CreateDrinkOrderDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDrinkOrderDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'coffeeBeanId': !exists(json, 'coffeeBeanId') ? undefined : json['coffeeBeanId'],
        'drink': !exists(json, 'drink') ? undefined : DrinkFromJSON(json['drink']),
        'milkType': !exists(json, 'milkType') ? undefined : MilkTypeFromJSON(json['milkType']),
        'isIced': !exists(json, 'isIced') ? undefined : json['isIced'],
    };
}

export function CreateDrinkOrderDtoToJSON(value?: CreateDrinkOrderDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'coffeeBeanId': value.coffeeBeanId,
        'drink': DrinkToJSON(value.drink),
        'milkType': MilkTypeToJSON(value.milkType),
        'isIced': value.isIced,
    };
}